requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
email-validator==2.0.0
python-dotenv==1.0.0